updated and fixed most unit tests
