+++ src/App.test.tsx
@@ -1,8 +1,13 @@
-import { render } from '@testing-library/react';
+import {render} from '@testing-library/react';
 import App from './App';
+import {FirebaseWrapper} from '@/components/FirebaseWrapper';
 
 test('renders without crashing', () => {
-  const { baseElement } = render(<App />);
+    const {baseElement} = render(
+	<FirebaseWrapper>
+	    <App />
+	</FirebaseWrapper>
+    );
   expect(baseElement).toBeDefined();
 });
