+++ src/components/FirebaseWrapper/FirebaseWrapper.test.tsx
@@ -1,0 +1,22 @@
+import {
+    expect,
+    it,
+} from 'vitest';
+import {
+    render,
+    screen
+} from '@testing-library/react';
+
+import {FirebaseWrapper} from '@/components/FirebaseWrapper';
+
+describe('FirebaseWrapper Component', () => {
+    test('should render with children', () => {
+	const element = render(
+	    <FirebaseWrapper>
+		<div data-testid='child'>
+		</div>
+	    </FirebaseWrapper>
+	);
+	expect(screen.getByTestId('child')).toBeDefined();
+    });
+});
