+++ src/components/Input/Input.test.tsx
@@ -3,7 +3,6 @@ import {
     afterEach,
     describe,
     expect,
-    it,
 } from 'vitest';
 import {
     cleanup,
