+++ src/pages/ResetPassword.test.tsx
@@ -3,6 +3,6 @@ import '@testing-library/jest-dom/extend-expect';
 import ResetPassword from './ResetPassword';
 
 test('renders Login page without crashing', () => {
-    const { baseElement } = render(<ResetPassword />);
+    const {baseElement} = render(<ResetPassword />);
     expect(baseElement).toBeDefined();
   });
